oci: spec: split conversion to oci/config/convert
The conversion code in oci/generate really belonged in a separate
package because it does not use any aspect of the image configuration
_generation_. Not to mention that conversion is actually going to be
defined in the spec, so it should be as self-contained as possible.
