+++ oci/config/convert/README.md
@@ -1,0 +1,11 @@
+### `umoci/oci/config/convert` ###
+
+One fairly important aspect of creating a runtime bundle is the configuration
+of the container. While an image configuration and runtime configuration are
+defined on different levels (images are far more platform agnostic than runtime
+bundles), conversion from an image to a runtime configuration *is* (or rather
+*will be*) defined as part of the OCI specification.
+
+This package implements a fairly unopinionated implementation of that
+conversion, allowing consumers to easily add their own extensions in the
+runtime configuration generation.
